backseat_app/main: Make runnable without modification
Many backseat app users will want to, y'know, actually use the backseat
interface and so write some amount of data processing code. But many
others just want to satisfy the vehicle's requirement for a heartbeat
and do their actual payload work elsewhere. Creating a repo using this
template should result in a runnable backseat application that publishes
a heartbeat. Remove the requirement to add a processing class before the
sample main is runnable.

While we're in here, fix a couple other things, like the name of the
default config file.